fix: update for imagegenerators framework
In the preview version of Zim (branch develop), there has been a change in the diagram editors:
https://github.com/zim-desktop-wiki/zim-desktop-wiki/commit/55ab6f2b69ee3e745bd267e56f11a53b7ad10aa3#diff-da5ad2e4e158ad4987acb4ba2616d16c439fc925fe822c1b0d71ab47ec571a19

Basically, the imagefile_extension declaration has moved to a different class.

This MR copies that change into the PlantUML plugin.

It hasn't landed on master yet, but when running from the Zim snapshot PPA, this change is needed (otherwise a bug occurs).